Refactor home.html and views.py
In home.html: revert form tag to not have action attribute (since
the home template should now only be called by the home_page view
and thus it is safe to POST back to the calling view).
Also remove the table as we will not be showing any existing items
with this template

In views.py we're no longer getting the items and passing them to the
template since the template will no longer show any existing items.